HiKoB Openlab release 2012.07
Improved agile-fox platform support.
Added support for dangerous-newt platform.
    
Improved peripheral support:
  rf2xx      : 802.15.4 compliant 2.4GHz radio module
  lsm303dlhc : 3D accelerometer + 3D magnetometer
  l3g4200d   : 3D gyroscope

new in this release

  lps331     : pressure and temperature sensors
  pga308     : programmable analog sensor signal conditioner
  pca9685    : 16-channel, 12-bit PWM

+++ lib/random/random.c
@@ -14,14 +14,14 @@
  * License along with HiKoB Openlab. If not, see
  * <http://www.gnu.org/licenses/>.
  *
- * Copyright (C) 2011 HiKoB.
+ * Copyright (C) 2011,2012 HiKoB.
  */
 
 /*
  * random.c
  *
  *  Created on: Sept 2011
- *      Author: Antoine Fraboulet <firstname.lastname@hikob.com>
+ *      Author: Antoine Fraboulet <antoine.fraboulet.at.hikob.com>
  */
 
 /*
@@ -45,6 +45,6 @@ unsigned short random_rand16(void)
 
 unsigned int   random_rand32(void)
 {
-    return (unsigned short)rand();
+    return (unsigned int)rand();
 }
